Change number type from Int to Long.

+++ app/src/main/java/com/example/numbers/data/dataSource/RetrofitDataSource.kt
@@ -9,7 +9,7 @@ class RetrofitDataSource(
     private val api: NumbersApi
 ) {
 
-    suspend fun getFactAboutNumber(number: Int): NetworkResult<FactResponse> {
+    suspend fun getFactAboutNumber(number: Long): NetworkResult<FactResponse> {
         return handleApi { api.getFactAboutNumber(number) }
     }
