+++ app/src/main/java/com/example/numbers/data/models/Fact.kt
@@ -2,7 +2,7 @@ package com.example.numbers.data.models
 
 data class Fact(
     val id: Long,
-    val number: Int,
+    val number: Long,
     val text: String,
     val found: Boolean
 )
